+++ Runtime/Core/Internal/IStore.cs
@@ -19,7 +19,6 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-using System.Collections.Generic;
 namespace Kingdox.UniFlux.Core.Internal
 {
     ///<summary>
@@ -27,10 +26,6 @@ namespace Kingdox.UniFlux.Core.Internal
     ///</summary>
     internal interface IStore<TKey, TStorage>
     {
-        ///<summary>
-        /// Store all the TStorage by TKey key
-        ///</summary>
-        IDictionary<TKey, TStorage> Storage {get;}
         ///<summary>
         /// Store TStorage with TKey
         ///</summary>